rework of the x11 module
Windows no longer have their private pixmap, instead
I only use one for double buffering.  This also implies
that the "window" abstraction is a lot dirtier than
before, so at least the naming of functions needs to be
reworked.

Also, I removed the useless width property of gui
fonts.

This is all nice, but we are still blinking!