fix tag display bug related to w->rev
The revision of the tag needs to be updated
when drawn.  I moved the revision bumping code
in draw() to handle both the tag and regular
windows uniformly.

It is unclear if this use of rev == 0 is sane
or not.  Indeed, the dirty(W *) code will update
the window only if rev != 0, this implies that
the following invariant holds:

!  Whenever rev == 0, the l[] array is  !
!  up-to-date with the buffer.          !

I do not know if this is true now.