superior resource management for async tasks
The previous implementation with reference counting on EBuf
was clunky.  First, use of negative reference counts made the
logic unclear.  Second, when an async task was started and a
file was reloaded in the window (eb_clr), some garbage could
be output to the new buffer at random positions.  Third,
killing a buffer could leak file descriptors tied to long
running tasks.

The problems are solved by eagerly destructing all the
resources of tasks that output in a given buffer right
before its destruction.